% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{find_graphviz}
\alias{find_graphviz}
\title{Returns a path to a graphviz file included in this package}
\usage{
find_graphviz(name)
}
\description{
Returns a path to a graphviz file included in this package
}

