% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_hours_per_team_plot}
\alias{make_hours_per_team_plot}
\title{Create a ggplot showing the discrepancy between calendar
hours and labor hours for diachronic and synchronic teams.}
\usage{
make_hours_per_team_plot(team_colors, line_size)
}
\description{
Create a ggplot showing the discrepancy between calendar
hours and labor hours for diachronic and synchronic teams.
}

