% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_time_data}
\alias{make_time_data}
\title{Make up some data used to show the time discrepancy between team structures.}
\usage{
make_time_data()
}
\description{
Make up some data used to show the time discrepancy between team structures.
}

