% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_time_plot}
\alias{make_time_plot}
\title{Make a base plot showing relationship between calendar and labor hours.}
\usage{
make_time_plot()
}
\description{
Make a base plot showing relationship between calendar and labor hours.
}

