% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{read_evoteams_chunks}
\alias{read_evoteams_chunks}
\title{Read all available knitr chunks in the evoteams package.}
\usage{
read_evoteams_chunks()
}
\description{
Read all available knitr chunks in the evoteams package.
}

