% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{time_interval}
\alias{time_interval}
\title{Compute time interval between first and last submissions.}
\usage{
time_interval(submissions)
}
\description{
Compute time interval between first and last submissions.
}

